Advanced payment

+++ database/migrations/2020_06_23_002202_create_tutor_levels_table.php
@@ -16,8 +16,8 @@ class CreateTutorLevelsTable extends Migration
         Schema::create('tutor_levels', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('tutor_level', 64);
-            $table->double('level_experience_lower_bound', 20,2);
-            $table->double('level_experience_upper_bound', 20,2);
+            $table->bigInteger('level_experience_lower_bound');
+            $table->bigInteger('level_experience_upper_bound');
             $table->double('bonus_rate', 10, 5);
         });
     }
